requirements.txt
Django
djangorestframework
django-autofixture
